almost good

+++ ft_bzero.c
@@ -6,7 +6,7 @@
 /*   By: stakaki <stakaki@student.42tokyo.j>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/04/09 21:42:49 by stakaki           #+#    #+#             */
-/*   Updated: 2021/04/12 13:28:36 by stakaki          ###   ########.fr       */
+/*   Updated: 2021/04/23 16:36:19 by stakaki          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 void	ft_bzero(void *s, size_t n)
 {
 	char	*cast_s;
-	int		i;
+	size_t	i;
 
 	cast_s = (char *)s;
 	i = 0;
